tests: cache virtualenvs, clean environment from python* variables

+++ tests/integration-tests/virtual_environments.py
@@ -1,6 +1,7 @@
 import os
 import shutil
 import subprocess
+import sys
 import virtualenv
 
 
@@ -20,16 +21,27 @@ def prepare_virtualenv(package_name=None, package_version=None):
     Prepares a virtual environment.
     :rtype : VirtualEnvDescription
     """
-    vdir = 'env'
-    if os.path.exists(vdir):
-        shutil.rmtree(vdir)
-    virtualenv.create_environment(vdir)
+    vroot = 'env'
+    env_key = get_env_key(package_name, package_version)
+    vdir = os.path.join(vroot, env_key)
+
     vbin = os.path.join(vdir, ('bin', 'Scripts')[_windows])
     exe_suffix = ("", ".exe")[_windows]
     vpython = os.path.join(vbin, 'python' + exe_suffix)
     vpip = os.path.join(vbin, 'pip' + exe_suffix)
+    venv_description = VirtualEnvDescription(home_dir=vdir, bin_dir=vbin, python=vpython, pip=vpip)
 
-    env = os.environ.copy()
+    # Cache environment
+    done_flag_file = os.path.join(vdir, "done")
+    if os.path.exists(done_flag_file):
+        return venv_description
+
+    if os.path.exists(vdir):
+        shutil.rmtree(vdir)
+
+    virtualenv.create_environment(vdir)
+
+    env = get_clean_system_environment()
     env['PIP_DOWNLOAD_CACHE'] = os.path.abspath(".pip-download-cache")
 
     if package_name is not None:
@@ -40,4 +52,22 @@ def prepare_virtualenv(package_name=None, package_version=None):
         subprocess.call([vpip, "install", package_spec], env=env)
 
     subprocess.call([vpython, "setup.py", "install"], env=env)
-    return VirtualEnvDescription(home_dir=vdir, bin_dir=vbin, python=vpython, pip=vpip)
+
+    open(done_flag_file, 'a').close()
+
+    return venv_description
+
+def get_env_key(package_name, package_version):
+    key = "%d.%d.%d" % (sys.version_info[0], sys.version_info[1], sys.version_info[2])
+    key += "-" + os.name
+    if package_name is not None:
+        key += "-" + package_name
+    if package_version is not None:
+        key += "-" + package_version
+    return key
+
+def get_clean_system_environment():
+    cur_env = os.environ.copy()
+    env = dict([(k, cur_env[k]) for k in cur_env.keys() if not k.lower().startswith("python")])
+
+    return env
